Moved Shader Logic In Rendering Engine

+++ src/core/Graphics.hpp
@@ -43,7 +43,7 @@ class Graphics
 };
 
 struct Character {
-    unsigned int TextureID; // ID handle of the glyph texture
+    float TextureID; // ID handle of the glyph texture
     glm::ivec2   Size;      // size of glyph
     glm::ivec2   Bearing;   // offset from baseline to left/top of glyph
     unsigned int Advance;   // horizontal offset to advance to next glyph
